requirements.txt
# 
# requirements
# 
#---package-------------version-----
    fabric
    pyyaml